final touches
